JCOR-208 Migrate to Play 2.6
This commit migrates the core and sample projects to Play 2.6.

The bulk of the changes are related to the removal of a few global
objects previously used in Play, mainly the BodyParsers and the
default ExecutionContext. These are now accessed via dependency
injection in classes extending BaseController (the successor to
Controller).

In addition, the ActionBuilder has two changes that impact
StackActionBuilder: it takes an additional type parameter
with the type of the request body, and it has an abstract method
parser through which the BodyParser is available.